Rename 'date-to-bytes.js' and 'bytes-to-date.js'.
