+++ modules/icao9303/utilities/date-to-bytes.js
